(CODEMGMT-1152) Implement path string for printing
When tracking content for purging, it's important for basemod modules
that their @path variable point to the actual root of the work-tree.
But, then when printing in logs, it would be hard to identify which
module was being deployed. So, implement a separate variable to use when
logging, so @path can be correct but logging can still make it obvious
which module is being worked on.