(maint) Pin CRI and note only Ruby >= 2.3 is supported
Previously, we had been floating the CRI version to be anything in 2.x
equal to, or above, 2.8. This has caused us to inadvertently drop
support for Ruby < 2.3. After some internal discussion we've decided to
stay with only supporting Ruby >= 2.3 and make our CRI version more
deterministic.

The rationale for the decision was that the PDK and Puppet 6 both ship
with Rubies that R10k 3 run on and those using Puppet 5 and earlier
should probably stay on R10K 2.x.