Implement R10K::Environment::WithModules baseclass
This commit implements a new capability for R10K environment objects:
environments can now contain modules directly, when built from an
appropriate source type and environment type.

Previously, modules could only be attached to an environment by loading
an environment, then reading the Puppetfile it contained. This
constrained the system because in order for modules to be added,
removed, or updated, either the core environment content needed to be
changed (the Puppetfile edited), or some kind of magic pattern needed to
be specified in the Puppetfile which delegated part of the work of
defining a module to an external source (the :control_branch pattern).

Now, the framework exists that enables an environment source to add,
remove, or update deployed module versions without ever touching or
modifying a Puppetfile.