Ensure that 'git clone' does not use hardlinks
Shell git uses hardlinks by default for local repo clones. This causes
issues with the r10k cache and SELinux contexts since the source and
destination will, quite often, be a home directory and some other
location on the system.

To prevent this (and other inadvertent issues with hardlinks in the r10k
cache), this patch adds the `--no-hardlinks` option to each call of `git
clone` when using shellgit.

Closes #931