Improve exception handling for threaded deploy
If a module thread raises a non-ThreadException, the behavior should
roughly match what serial deployment does. The exception should cause
r10k to halt the deploy.

This commit implements that behavior, while also ensuring that we don't
interrupt in-progress threads that have not raised an exception
themselves. Instead we wait for threads to finish their current module
before they halt and the main thread re-raises the first thread
exception encountered.