(CDPE-1653) Implement default-branch-override flag to override Puppetfile :default_branch entry during environment deployments.
With this change, a user can specify a default branch override using the `--default-branch-override` flag. This value will be passed through to the deployment and used in place of the Puppetfile's `:default_branch` entry. If no `:default_branch` was previously specified in the Puppetfile, the supplied default branch override will be used anyways.