(maint) Consistently compare values to nil
`overrides[:deploy]` is initialized as an empty Hash if any of
puppet-path and generate-types is set.

`overrides[:deploy][:puppet_path]` and
`overrides[:deploy][:generate_types]` are respectively set when
puppet-path and generate-types are non-nil.

`false` being not nil, when generate-types is set to `false`,
`overrides[:deploy]` will not be initialized and an excpetion will be
raised when setting `overrides[:deploy][:generate_types]` to false.

Always compare values to nil to avoid such a situation.