Extend Git environment type using WithModules
This commit changes the R10K::Environment::Git class's parent from
R10K::Environment::Base to  R10K::Environment::WithModules. The
practical effect of this is that Git environments are now compatible
with environment sources that directly define modules to deploy (in
addition to or instead of modules defined in a Puppetfile).