Bugfix: version absent on install for Forge mods
Prior to this fix, the .r10k-deploy.json file wouldn't have a module
version in it on first install. It would be present after an upgrade or
other modification of an already-present module, just not the first time
the module was added to an environment.

After this commit, the .r10k-deploy.json file is correctly populated
with new-install Forge modules' version numbers.