Add Yaml environment source (#983)
* Implement R10K::Environment::WithModules baseclass

This commit implements a new capability for R10K environment objects:
environments can now contain modules directly, when built from an
appropriate source type and environment type.

Previously, modules could only be attached to an environment by loading
an environment, then reading the Puppetfile it contained. This
constrained the system because in order for modules to be added,
removed, or updated, either the core environment content needed to be
changed (the Puppetfile edited), or some kind of magic pattern needed to
be specified in the Puppetfile which delegated part of the work of
defining a module to an external source (the :control_branch pattern).

Now, the framework exists that enables an environment source to add,
remove, or update deployed module versions without ever touching or
modifying a Puppetfile.

* Implement R10K::Source::Hash abstract baseclass

The Hash environment source type is an abstract base class that makes it
easy to define environments by simply enumerating them. For example, a
REST API could provide a JSON object enumerating each environment that
should exist, along with details such as modules to deploy to them. The
Hash source baseclass can consume this data and produce the appropriate
environments.

Note that such a source would no longer need be tightly coupled with
version control branches, which is the only way today's r10k sources
work.

* Extend Git environment type using WithModules

This commit changes the R10K::Environment::Git class's parent from
R10K::Environment::Base to  R10K::Environment::WithModules. The
practical effect of this is that Git environments are now compatible
with environment sources that directly define modules to deploy (in
addition to or instead of modules defined in a Puppetfile).

* Bugfix: version absent on install for Forge mods

Prior to this fix, the .r10k-deploy.json file wouldn't have a module
version in it on first install. It would be present after an upgrade or
other modification of an already-present module, just not the first time
the module was added to an environment.

After this commit, the .r10k-deploy.json file is correctly populated
with new-install Forge modules' version numbers.

* Implement yaml source type

This allows r10k to read environment definitions from a YAML file,
located by default at /etc/puppetlabs/r10k/environments.yaml.

* Fix error relating to purging environments w/ mods

* Implement "bare" environment type

This type has no content besides environment-attached modules

* Add experimental documentation for new features

Yaml source
Environment modules
Bare environment type