(RK-360) Allow overriding the cachedir in `deploy module`
Previously, we added the ability to configure the cachedir from the CLI
for the `deploy environment` command, since Code Manager configures a
separate cachedir for each of its r10k workers. Now that we are also
using the `deploy module` command in Code Manager, it needs to respect
these special cachedirs the same way. The `--cachedir` flag can now be
used with the `deploy module` command.