(maint) Only release from master
This commit removes 2.6.x as a branch to trigger the Tag and Release Action.
The action we use to tag always looks for the latest (semver) tag, regardless
of the branch that triggered the action. This would result in 3.x tags on the
2.6.x branch. Since we expect all community contributions to land in the master
branch, we can use our old, jenkins-based release process for maintenance
branches.