Improve hash source type
Two main improvements being made:

 - Include logging in hash source type
 - Split hash source #initialize from #environments
 - Add validation method for checking environment hashes

Logging is being added so that derived classes can use it.

Splitting the #initialize logic pertaining to creating environments and
and sending it into the #environments method instead allows child
classes to better control when (or if) side-effect actions are
triggered.

The validation method can be used to check if an environments hash is
valid. It only performs perfunctory validation today.