(maint) Remove second non_module_object from test
After increasing the default pool_size, a bug surfaced when attempting to
install multiple objects from the same git source. We attempt to cache both
objects, but the second attempt fails because we try to clone to the same
destination. We don't think this is common behavior and this test does not need
two objects to test what it's testing.