(maint) Remove CHANGELOG note re default thread pool size
After updating the default pool_size, a race condition was
found when initializing our local git cache. See
https://github.com/puppetlabs/r10k/issues/1058 for the issue in detail.

When attempting a fix we realized that there were multiple copies of the
code to iterate over modules to manage. See
https://github.com/puppetlabs/r10k/pull/1059 for a first attempt at this
problem and discussion.

In order to meet our current timelines, the default will need to reverted
until we can fix this issue properly. Hence this commit and a55536cad.

This commit was necessary because the CHANGELOG has been refactored since
the commit referencing the default was added.