(RK-319) Ensure r10k cleans up tmp directories
Prior to this commit, deploying modules from the forge would leave
behind tmp directories. The conditional check looked for the existance
of the download and extract paths, but not for the tmp directory
containing them. This commit cleans up the tmp directory instead of the
specific download or cache path.