[#1058] Add acceptance test for repeated git remote
This commit adds a beaker acceptance test verifying that it is possible
to use the same remote for several non-module objects. This case
previously broke when the `pool_size` feature was introduced, because
there was a race condition when creating the git cache in shellgit mode,
when the same remote was used in two objects being set up concurrently.