Revert "Issue 1096: use Default-Branch"
This reverts commit 57329b664e4aee23d29d55142b7f620489879df0.
This constituted a breaking change, which we are not prepared to make right now.
In the future, when we are ready for a new major version, we hope to require
users to specify a ref in their Puppetfile, rather than defaulting to something.
See https://github.com/puppetlabs/r10k/issues/1110 for more info.