(maint) Replace deprecated Cri `required` calls
Cri has deprecated the use of `required` to denote an option with a
required arugment because it was too confusing (it does not indicate
that the option itself is required, just its arg). It is now preferred
to use `option` with the `argument: :required` keyword arg. This commit
replaces our instances of `required` with the preferred form.