[#1106, #1109] Save credentials from CLI flags
This commit adds a step to the `deploy environment` command invocation
that takes the credential files as passed via the new flags, and passes
the relevant info through to the deployment object, which is responsible
for performing the bulk of the action logic.

For SSH keys, the file path is passed directly, as Rugged expects to be
passed a file path, not the key itself. Like the rest of r10k, this
still does not support password-protected keyfiles.

For OAuth, the contents of the file is read and passed through, since
Rugged expects to be given the token as a string. For `token-path`, we
also support reading from stdin when the conventional `-` is passed.