[#1106, #1109] Plumb git creds from CLI to rugged
This commit uses r10k's settings architecture to supply git credentials
passed via command line options to the Rugged credentials object. It
adds a setting for `oauth-token` that mirrors `private-key` and can be
used to specify a token file either globally or per-repo in the config,
as well as passed on the command line. Credentials passed on the command
line have the highest precedence, followed by per-repo values from the
config, before falling back to the global values.