Implement standard module option schema
Allow modules of any currently supported type to be initialized using
the following standard specification:

  ---
  <module-name>:
    type: <type>
    source: <source>
    version: <version>

This is in slight contrast to today's expectations, which are largely
based on the presence of a type-specific key to indicate the kind of
module to create, and there are no standard parameters to indicate
source and version.

By providing a standard schema for defining modules of any type, this
paves the way for easier integration with external data systems which
may define sets of modules of many types to include in different
environments.

This schema is generalized away from implementation and towards the
concepts that matter to the user.