(maint) Base new tags on the current branch
Previously, the GitHub Action to tag a new release would base the new tag on the
latest tag in the repo, rather than on the latest tag for the *current branch*
of the repo. This change allows us to release from various branches without
worrying about newer tags landing on older branches.