(maint) Deprecate --puppetfile for deploy display
When displaying a deployment's information we would accept a
'-p' or '--puppetfile' to display information about modules attached
to the environment via a Puppetfile, yaml[dir], or exec. This was
originally a nod to the fact that most modules were vendered within
the environment and the Puppetfile was at that time the only way to
attach modules to an environment.

Since then, modules may be attached to an environment via non-Puppetfile
means, and attaching modules (rather than vendoring them) has become
the preferred means of managing them.

This deprecates the '-p' and '--puppetfile' flags for '-m' and '--modules'
as that more closely follows current best practices.