(maint) Do not raise NoMethodError on normal env info writing
Previously when calling write_environment_info we would call
Module#repo to collect the sha, if it existed. However, only git backed
modules have this method defined. So the write_environment_info method
would cause an exception which it caught and ignored when inspecting any
non-git-backed modules.

This moves to use the Module#properties method which is well defined for
every type of module and has an `:actual` key which matches the value
that the write_environment_info used to gather. We do not collect
`:actual` key values for any module besides those of type `:git` to
match the behavior preivously.

This also adds a test for the failure case of the write_environment_info
method that was previously untested. I have a feeling that this is no
longer needed now that we do not call Module#repo indiscriminantly,
however it is included to err on the safe side.