(maint) Use a "Cleaner" helper to purge Puppetfile content
Prior to this Puppetfile content was managed via the Purgeable mixin.
As a long term direction it would be valuable to move the Puppetfile in
a direction of acting on, or contributing to, modules connected to an
environment. In the processing of "dumbing" down the Puppetfile class
this patch moves the responsibility of purging to a helper class
(`R10K::Util::Cleaner`), leaving the Puppetfile class only the
responsibility of determining the Cleaner's inputs.