(maint) Use Ruby 2.3's Hash#dig & default arg values
R10K only supports Ruby >= 2.3 which means we can use Hash#dig instead
of Hash#fetch#fetch.

Also, it is never okay (outside of testing) to not pass a Hash for
the Action#new `settings` arg. Because of this, make the setting
argument mandatory.

for the settings, no need for a `nil` default that we later
conditionally override.