(maint) Put all deploy modules configuration into @settings
Ideally, much of the logic in visit_* callbacks would live in the
objects (Deployment, Environment, etc) rather than the Action. As a
first step we put all of the configuration that is used in used into
the @settings instance variable that we pass into the Deployment.
This gives us the option to further pass the values into child
objects of the Deployment. When _actually_ using this configuration
by implementation objects I assume a better structure will be found
and some variables may be pulled back out to stay only in the
Action's scope. This should be considered a naive first pass across
the Actions.