(maint) Move control branch warning to git module init
Prior to this when running `puppetfile install` users would receive a
warning whenever processing a git module with a control_branch setting
of :control_repo. This is because the puppetfile subcommands have no
concept of a containing environment that they operate within.

However, this warning should also be applicable when deploying git
modules with non-git backed environments (eg SVN or "bare"
environments). This warning is also only applicable to git backed
modules.

This patch moves the warning from the `puppetfile install` command to
the git module initialization and changes the message to be more general.