(maint) Move type generation for `depoy module` to #visit_environment
For every module deployed via `deploy module`, if the `generate-types`
flag was given, we would attempt to generate types for the environments
containing that module. This could lead to type generation happening
multiple times for each environment, though in practice users typically
use this action to update a single module at a time.

This moves the logic regarding type generation from the visit_module
method on the deploy modules action to the visit_environment method
where it makes more sense to do environment wide updates.