(maint) Remove visit_module usage
Previously, the module.sync method would be called within a visit_module
method on each action controlling the sync.

This removes the visit_module methods from actions and has the locations
where `module.accept(visitor)` as called before call `module.sync`
instead.

The visitor interface would wrap every visit_* method in a rescue block
and set the `visit_ok` instance variable to false if there was an issue.
To keep this general behavior for each action we've add this behavior to
the `call` method of the action.