(maint) Provide Visitorless means of syncing Puppetfile
Previously to sync the content of a Puppetfile a caller had to implement
the Visitor pattern. This commit provides a `sync` method on the
Puppetfile object that does not require a visitor. It also refactors the
ContentSynchronizer to offer `concurrent_sync` and `serial_sync`
versions of its `concurrent_visit` and `serial_visit` methods.