(maint) Move puppetfile purge info to module loader
This moves the information gathering required to purge at the Puppetfile
level (which is a collection of module info) into the
ModuleLoader::Puppetfile. The existing Puppetfile logic defers to the
new puppetfile module loader. Several tests are commented out because
stubbing the information from the module loader is difficult from the
puppetfile test. These tests are commented out to provide a guide for
the next commit, which will flesh out testing of the puppetfile module
loader.