(maint) Add tests for Puppetfile module loader
This commit audits the R10K::Puppetfile tests and moves/rewrites those
that are applicable to the R10K::ModuleLoader::Puppetfile tests. It also
makes some refactors to both for easier testing.

One notable behavior change in R10K::Puppetfile is that the `load`
method will now return the loaded content (which is truthy) rather than
the boolean true.