(CODEMGMT-1448) Reload environments before deployment purges
When using git deploy pools of r10k, new branches can be created and deployed
that won't exist in the current downloaded cache; during a deployment
purge, this would end up deleting an in-process, simultaneous
deployment of another branch. This change adds an interface for a source
to reload its environments mid-deploy; that `reload!` method will be called
before running deployment purges to help ensure that the list of
environments is up to date.