(maint) Simplify forge module implementation IDing
Anywhere except the Puppetfile, a module input must specify
{type: forge} to indicate a Forge module. In the Puppetfile, we convert
a scalar version arg to a hash already, before passing it to
R10K::Module.new(). So, let's simplify the implementation selection
logic.

In the Puppetfile, convert to Hash with both both version: and type:
keys, when given scalar module args, which by convention we know to be
Forge modules.