(maint) Re-allow relative and absolute paths for moduledir & puppetfile
Previous to recent refactors an API user could provide either an
absolute or relative moduledir path, as well as an absolute
puppetfile_path or relative puppetfile_name, when resolving Puppetfiles.
The recent refactor required consumers to either pass absolute paths or
allow defaults rooted in the basedir, but disallowed passing relative
paths that would then be rooted in the basedir.

This commit returns that ability, as well as refactoring the internal
path joining and cleaning to be shared by all consumers of a file system
path.

This will change the the internal storage of basedir, moduledir, and
puppetfile to use a "cleaned" path stripped of extra slashes and dots.
However, this cleaned representation is what we would compare to
install_paths and display to users in error messages, so it should be a
change that does not affect any behavior.