(maint) Rework default_branch_override passing for Puppetfile
Previously, the Puppetfile ModuleLoader would not extract and use the
default_branch_override from its options, which would cause the its
usage to fail.

We also would extract the wrong key for the default_override_branch in
R10K::Puppetfile, causing us always warn that the passed override and
the initialized one was wrong, even when they matched.

This patch also includes some minor fixes to how defaults are handled
and local variables are named.