Add support for Github App auth token
This authentication benefits large organizations with strict SSO/2FA
guidelines that cannot utilize machine users for puppetserver
deployment.

Here are some of the advantages with Github App vs OAuth or SSH:

* The rate limit of server-to-server requests scales with user + repo count
* No seat consumption with Github App vs machine user. Enforcing SSO/2FA on
  the organization restricts machine user capabilities
* No need to manage unique ssh keys per repository for deployment, which
  can grow upward into the thousands with multiple puppetservers

It has three parameters, one of which is optional with a default
value:

github-app-id: Github App Identifier
github-app-key: SSL private key
github-app-ttl: Token time-to-live in seconds (optional, default: 120)

It generates a new token for every repo when calling `environment`
subcommand.

The jwt gem was added to the gemspec config in order to create the
json web tokens requests.