(maint) Fix broken acceptance test
Previously, the single_env_purge_unmanaged_modules test said that
it was testing that agent runs would fail because the motd module
had been successfully purged.

However, it was actually testing that a) the purge command failed and
b) the agent run succeeded but the assertion failed, not because of
anything that happened on SUT, but because it was testing the output
against the regex /Blah/ which should not appear in normal agent output.
The test then claimed that the failure was part of CODEMGMT-78, which
was a feature improvement to `r10k puppetfile` to respect `--moduledir`
and `--puppetfile` flags. However, that feature was implemented in 2016
and the test has continued to "pass" because of a failing purge command
because the test never actually attempted to use either the moduledir
or puppetfile flags.

This patch updates the purge invocation to actually use moduledir and
puppetfile flags, updates the expectation on the purge invocation to
be successful. It also updates the expectation on agent runs to be
unsuccessful but removes the duplicitous expectation on output matching
/Blah/.