(maint) Update default OS for integration tests
Previously we tested on CentOS 6 which is no longer a supported platform
on our internal tooling.

This patch continues to use CentOS since it is the traditional platform
for our FOSS project, though due to some concerns about the direction of
CentOS 8 we stay with CentOS 7. When we must move off of it, there will
hopefully be more guidance as to what Enterprise Linux variant to use in
FOSS testing.