(maint) Raise when calling Puppetfile#load! if file is missing
As part of a recent refactor the conditional behavior to check if a
Puppetfile existed was moved from `load` to `load!`. However, the
puppetfile actions only call `load!` and depend on its behavior of
raising for parts of their user experience.

This commit moves the check for Puppetfile existence back to `load` so
that it can be skipped by the puppetfile actions and they will properly
error for users when ran in a directory that does not contain a
Puppetfile.

It adds tests for the `load!` method that exercises its lack of
idempotence. It also removes the `freeze` from remaining internal module
loader objects as the method will still fail, but fails with a more user
friendly error (R10K::Error, "duplicate content") rather than
FreezeError.