(CODEMGMT-1421) Plumb option for deploy_spec
This commit introduces no functional change; it just plumbs through
an option for `deploy-spec` to be set at either the CLI or in the
puppetfile. If passed via the CLI or in r10k's config, that will
take global precedence for all modules. If not specified globally,
individual modules can specify this specifically in each module's
definition in a Puppetfile.

The deploy_spec option defaults to false; future work will actually
implement the option for relevant R10K::Module subclasses.