(RK-369) Run postrun command for module deploys
This commit updates the `deploy module` command to run the configured
`postrun` command that is normally run after an environment deploy. It
will only run the command if any of the requested environments were
actually modified, i.e. if they contained any of the modules we asked to
deploy. Similarly, if the command contains the `$modifiedenvs` variable,
it will be populated only with environments that actually had modules
deployed. This mirrors how we run `puppet generate types` for module
deploys when that setting is enabled.