(CODEMGMT-1415) Allow assuming modules are unchanged between deploys
This commit adds a command line flag `--assume-unchanged` that will
cause the Environment to load the existing Puppetfile when it is created
(prior to having updated the backing store if backed by a VCS). It loads
this original Puppetfile instructing the DSL to only load module
Definitons rather than modules backed by sync-able implementations
(Forge, Git, etc).

Definition modules (R10K::Module::Definition) determine their version
without doing further I/O (only the definition in the Puppetfile is used)
and implements `sync` as a no-op.

From these, module definition name/version pairs are created and stored
by the normal Puppetfile module loader which it checks when it loads the
updated Puppetfile later on. If a module name/version pair exists, the
version is non-nil, and matches the updated Puppetfile then a module
Defnition is placed in the modules queue to be synced, if these criteria
are not met then the module Definition is converted into its proper
implementation class for syncing.

As previously mentioned the queued module Definition implements `sync`
as a no-op.

Note: this commit also clarifies that Module::Base.new requires the
"opts"/"args" parameter to be a Hash rather than an Array (as all
subclasses have required for some time).