(CODEMGMT-1454) Ensure missing repo caches are re-synced
This commit updates the `status` method for stateful git repos to check for the
presence of the git cache when determining whether repo updates are necessary.
Previously, if the repo's working dir and git dir were present, but the cache
was missing, the `@repo.dirty?` check would fail when attempting to diff the
workdir (`object not found - no match for id`). This would occur during an
environment deploy if the environment used the same remote repo as another
environment that had been deleted (and thus its artifacts cleaned up). By
setting the status to `:mismatched` when the cache doesn't exist, we ensure that
the cache is resynced and the deployment may proceed.