(CODEMGMT-1415) Ensure environment purge_exclusions are calculated
In commit 9d0579f07 Environments stopped interrogating their Puppetfile
for managed content needed to calculate their purge exclusions, but
moved to using the data returned by the ModuleLoader when loading
content.

However, if content was not loaded prior to attempting to purge the
environment then the environment would purge much of the Puppetfile
managed content. This would happen when running `r10k deploy
environment` (without `--modules`) on an existing environment.

To resolve this we now conditionally load module data if it hasn't been
loaded yet when resolving our purge exclusions.