(RK-308) Allow respecting `forge` declaration in puppetfile
This commit adds a new setting `forge.allow_puppetfile_override`, that
when true will cause a `forge` declaration in a Puppetfile to override
the `forge.baseurl` setting for contacting the Forge. This allows users
to configure per-environment Forge URLs, for example, if an internal
Forge is used for a test environment and the live Forge for production.
Previously, a user would have had to choose one or the other. The
setting is false by default, since many users have erroneous `forge`
declarations in their Puppetfiles that were previously completely
ignored.