(maint) Remove unused `forge` instance var from ModuleLoader::Puppetfile
This class does not make any use of its `forge` instance variable, nor
does it expose it via a getter. Since it is unused, this commit removes
it and the associated tests.

Note that the old `Puppetfile` class still exposes a `forge` instance
variable that is hard-coded to a default, and does not respect either
the `forge.baseurl` setting or a `forge` declaration in a Puppetfile.
This var is also unused within r10k (`PuppetForge.host` is set directly
elsewhere), but might be used by people using the `Puppetfile` class by
itself.