[#1215] Ensure module postrun command only runs for modified envs
Previously, when a postrun command with interpolated environments was
specified, the `deploy module` command with no environments specified
would cause the command to be run for all environments that contained
the module, whether or not the env was actually updated by the
deploy.

Similarly, `puppet generate types` would be run for any environment that
contained the module, not just those for which it was updated. Together,
these two things could result in a huge amount of unnecessary work.

This commit makes modules return the result of their syncing, true for
"updated" and false for "not updated". The environment's `deploy`
function collects this information, recording the names of each module
that was updated as part of the deploy. The `module deploy` action then
uses this to determine which environments actually had modules modified,
and uses that to populate the `$modifiedenvs` in the postrun command. So
now, the command will only be run for environments that actually updated
a module.

This also updates the logic for running `puppet generate types` to
follow the same pattern: it is only run for environments that actually
had an updated module.

Note this does not affect environment deploys.