Purge should purge directories too, not just files
Previously, purge! would never remove directories. This awkwardly left
behind large file-less directory trees, when e.g. a module was removed,
or a directory stopped being managed in an environment.

For the latter use case, this was probably hidden somewhat by Git doing
empty directory cleanup on checkout. That won't always be guaranteed to
happen, when not using Git sources.