Adds ability to resolve dependencies into a lockfile
This modifies the path resolution so that /path/to/Puppetfile.lock is
preferred. If the lockfile exists, then it will be used instead of the
Puppetfile. It also adds a `puppetfile resolve` action that will resolve
dependencies and write out that lockfile.

Note: all dependencies will be satisfied from the Forge, no matter what
the original source was.

Expected user workflow:

* Write Puppetfile that describes only the modules you intend to use.
* Run `r10k puppetfile resolve` to resolve all dependencies of those
  modules.
* Optional: review the dependent modules for quality and security
  purposes.
* Deploy.

@glennsarti what do you think about this implementation? It feels a bit
weird to have multiple Puppetfile parsers in the same tool, but don't
really see a better way. And am I using the dependency graph properly?