Align terminology with Puppet; bare => plain (#1228)
Align terminology with Puppet; bare => plain

Puppet already uses a term in the indirector system to indicate that an
implementation doesn't really do anything, and that term is "plain". For
the R10K environment type that doesn't do anything, we should use this
same terminology. The initial choice of the term "bare" isn't ecosystem
congruent, and has the language problem of being used meaningfully when
describing Git repos, but that meaning doesn't translate at all to the
environment type.

Using the term "plain" both aligns terminology with Puppet, and
eliminates any potential language collisions with the Git meaning of the
term.