(CODEMGMT-1294) Resync repos with unresolvable refs
Prior to this commit, if the git repo did not have a valid HEAD or the
ref was not resolvable, r10k would raise an error and fail. This commit
changes the behavior to do a forceful resync of the git repo.
This commit updates the `status` method for stateful git repos to return
:mismatched when the repo.head is nil. This commit updates the resolve
method for rugged base repositories to catch `Rugged::OdbError`, which
would have resulted in `object not found - no match for id` messages.
This condition would occur when the local repo was out of sync with the
upstream repo and a local commit could not be found in the upstream
repo. Since this condition indicates that the repo has diverged from the
upstream repository, this change forces a full resync of the git repo to
ensure it is in sync.